updated logic to use BigDecimal instead of double

+++ app/src/main/java/com/example/budgetplanner/Purchase.kt
@@ -1,8 +1,11 @@
 package com.example.budgetplanner
 
+import java.math.BigDecimal
+import java.text.DecimalFormat
+
 class Purchase (
     val vendor: String,
-    val price: Double
+    val price: BigDecimal
     ) {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
